requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchain-chroma
sentence-transformers
scikit-learn
chromadb
